requirements.txt
litellm
numpy
pandas
scipy
pydantic
pillow

